Factor ContentFilterStats out into its own file
This will be used by TabImpl.
